+++ functions/handlers/emailHandler.js
