+++ functions/handlers/pdfHandler.js
